test: add full-coverage tests and fix Newtonsoft converter date-string bug
Agent-Logs-Url: https://github.com/porrey/UnixTime/sessions/8d469655-470a-46fd-8f79-773e6b3841a2

+++ Src/System.UnixTime.Tests/UnixTimeTests.cs
@@ -55,7 +55,8 @@ namespace System.Tests
 		[TestCaseSource(nameof(Items))]
 		public void IntConstructorTest(TestDataItem data)
 		{
-			UnixTime target = new(data.UnixTimestampInt);
+			// Cast explicitly to int so the int overload is exercised, not the long overload.
+			UnixTime target = new((int)data.UnixTimestampInt);
 			Assert.That(target.DateTime.ToUniversalTime(), Is.EqualTo(data.DateTime));
 		}
